test(proof): verify Denim range in SP1
Amp-Thread-ID: https://ampcode.com/threads/T-01a0214a-f554-75e9-a3aa-d57ff8ad375f
