+++ .github/workflows/denim-sp1.yml
@@ -1,0 +1,55 @@
+name: Denim SP1
+
+on:
+  pull_request:
+    paths:
+      - ".github/actions/setup/action.yml"
+      - ".github/workflows/denim-sp1.yml"
+      - "Cargo.lock"
+      - "Cargo.toml"
+      - "crates/batcher/comp/**"
+      - "crates/common/**"
+      - "crates/consensus/**"
+      - "crates/proof/**"
+      - "etc/just/succinct.just"
+  push:
+    branches: [main]
+    paths:
+      - ".github/actions/setup/action.yml"
+      - ".github/workflows/denim-sp1.yml"
+      - "Cargo.lock"
+      - "Cargo.toml"
+      - "crates/batcher/comp/**"
+      - "crates/common/**"
+      - "crates/consensus/**"
+      - "crates/proof/**"
+      - "etc/just/succinct.just"
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.head_ref || github.run_id }}
+  cancel-in-progress: true
+
+permissions:
+  contents: read
+
+jobs:
+  denim-sp1:
+    name: Execute Denim range ELF
+    runs-on:
+      group: BasePerfRunnerGroup
+    timeout-minutes: 60
+    steps:
+      - name: Harden the runner (Audit all outbound calls)
+        uses: step-security/harden-runner@9af89fc71515a100421586dfdb3dc9c984fbf411 # v2.19.4
+        with:
+          egress-policy: audit
+      - uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
+        with:
+          fetch-depth: 1
+      - uses: ./.github/actions/setup
+        with:
+          free-disk: "true"
+          native-deps: "true"
+          sp1: "true"
+          rust-cache-shared-key: "denim-sp1"
+      - run: just succinct test-denim-range
