+++ crates/proof/zk/utils/src/test_utils/mod.rs
@@ -1,0 +1,6 @@
+//! Reusable deterministic fixtures for proof-backend tests.
+
+mod denim;
+pub use denim::{
+    CLAIM_BLOCK, DENIM_FIXTURE_CONTENT_HASH, DENIM_TIMESTAMP, DenimFixture, ExpectedDenimBlock,
+};
