+++ etc/just/succinct.just
@@ -20,8 +20,13 @@ build-range-elf:
         --elf-name range-elf-embedded \
         --tag {{ _sp1_tag }} \
         --docker \
+        --ignore-rust-version \
         --workspace-directory {{ _repo_root }} \
         --output-directory {{ _succinct_dir }}/elf
+
+# Execute the deterministic Denim fixture in the real local SP1 range ELF.
+test-denim-range: build-range-elf
+    cargo test --locked -p base-proof-zk-backend --test denim_sp1 -- --ignored
 
 # Build the aggregation program ELF (reproducible docker build)
 build-agg-elf:
@@ -33,6 +38,7 @@ build-agg-elf:
         --elf-name aggregation-elf \
         --tag {{ _sp1_tag }} \
         --docker \
+        --ignore-rust-version \
         --workspace-directory {{ _repo_root }} \
         --output-directory {{ _succinct_dir }}/elf
